functions/contact-form/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
flask==3.*
